command: handle all actual work in separate single thread
With the addition of Spotify integration, we're starting to regularly
exceed Slack's three second limit for responses from slash commands.
Rocket's State<T> is difficult to move into new threads, so instead of
directly spawning a thread to run `command_np` in the /np route, start a
dedicated thread for doing the actual work of responding to /np commands
and simply pass in requests via a Sync channel.